refactor: 토너먼트 아이템 타입 정의를 공통 타입으로 분리
