+++ apps/web/src/app/tournament/[id]/create/_types/tournament.ts
@@ -1,19 +1,13 @@
 import type { TOURNAMENT_STATUS } from '@/consts/tournament';
+import type { TournamentItemT } from '@/types/tournament';
+
+export type { TournamentItemT };
 
 export type TournamentItemLinkDataT = {
   itemId: number;
 };
 
 export type TournamentStatusT = (typeof TOURNAMENT_STATUS)[keyof typeof TOURNAMENT_STATUS];
-
-export type TournamentItemT = {
-  tournamentItemId: number;
-  itemId: number;
-  name: string;
-  price: number;
-  currency: string;
-  imageUrl: string;
-};
 
 export type TournamentParticipantT = {
   userId: string;
