+++ apps/web/src/types/tournament.ts
@@ -1,0 +1,8 @@
+export type TournamentItemT = {
+  tournamentItemId: number;
+  itemId: number;
+  name: string;
+  price: number;
+  currency: string;
+  imageUrl: string;
+};
